Move all the rendering  code to the builtin function struct
